Move firebase routes into separate file
